refactors: test configuration for PostgreSQL
Centralizes PostgreSQL Testcontainer management using a
singleton pattern and a base integration test class.

This streamlines test setup, avoids redundant container
initialization, and improves test suite performance by
reusing a single container instance across multiple tests.

Removes the need for `@Testcontainers` and
`@DynamicPropertySource` annotations in individual test
classes, simplifying their configuration.
